requirements.txt
flask>=2.0
python-dotenv>=1.0
torch>=2.0
torchvision>=0.15
opencv-python-headless>=4.5
Pillow>=9.0
